requirements.txt
absl-py==0.10.0
acme==1.1.0
altair==4.1.0
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
astroid==2.4.1
astunparse==1.6.3
atomicwrites==1.1.5
attrs==19.3.0
autopep8==1.5.7
backcall==0.2.0
base58==2.0.1
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
boto3==1.14.57
botocore==1.17.57
Brlapi==0.7.0
cachetools==4.1.1
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
comtypes==1.1.7
ConfigArgParse==0.13.0
configobj==5.0.6
configparser==5.0.0
cryptography==36.0.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.5.0
distro-info===0.23ubuntu1
Django==3.1.7
django-storages==1.11.1
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
duplicity==0.8.12.0
entrypoints==0.3
enum-compat==0.0.3
facebook-sdk==3.1.0
fasteners==0.14.1
filelock==3.0.12
fire==0.3.1
Flask==1.1.2
Flask-Cors==3.0.10
frappe-bench==5.7.5
future==0.18.2
gast==0.3.3
gitdb2==2.0.6
GitPython==2.1.15
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-colab==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.31.0
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
honcho==1.1.0
httplib2==0.20.2
idna==2.8
importlib-metadata==1.5.0
imutils==0.5.4
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.14.0
josepy==1.2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mysql-connector-python==8.0.28
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
nose==1.3.7
notebook==5.2.2
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.4.0.42
opt-einsum==3.3.0
packaging==20.3
pandas==0.24.2
pandocfilters==1.4.2
paramiko==2.6.0
parsedatetime==2.4
parso==0.7.1
pathtools==0.1.2
pbr==5.5.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==11.9.0
pluggy==0.13.0
portpicker==1.2.0
prometheus-client==0.8.0
prompt-toolkit==1.0.18
protobuf==3.13.0
psutil==5.5.1
ptyprocess==0.6.0
py==1.8.1
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycups==1.9.73
pydeck==0.5.0b1
Pygments==2.6.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pylint==2.5.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.17.1
PySocks==1.7.1
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.6
python-crontab==2.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.18.0
pyttsx3==2.90
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
reportlab==3.5.34
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.3.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
semantic-version==2.8.5
Send2Trash==1.5.0
setuptools-rust==1.1.2
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
smmap==5.0.0
smmap2==3.0.1
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
ssh-import-id==5.10
stevedore==3.2.1
systemd-python==234
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
testresources==2.0.0
texttable==1.6.3
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==4.5.3
traceback2==1.4.0
traitlets==5.0.4
tweepy==3.10.0
typing_extensions==4.0.1
tzlocal==2.1
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
uritemplate==4.1.1
urllib3==1.24.3
usb-creator==0.3.7
validators==0.18.1
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
watchdog==0.10.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
